+++ Carmu/Sources/Views/MyPage/CrewInfo/CrewInfoNoCrewView.swift
@@ -9,19 +9,11 @@ import UIKit
 
 final class CrewInfoNoCrewView: UIView {
 
-    private let contentView: UIView = {
-        let contentView = UIView()
-        contentView.layer.cornerRadius = 20
-        contentView.layer.masksToBounds = true
-        contentView.backgroundColor = UIColor.semantic.backgroundDefault
-        return contentView
-    }()
-
     private let noCrewStackView: UIStackView = {
         let noCrewStackView = UIStackView()
         noCrewStackView.axis = .vertical
         noCrewStackView.spacing = 20
-//        noCrewStackView.alignment = .center
+        noCrewStackView.alignment = .center
         return noCrewStackView
     }()
 
@@ -37,15 +29,20 @@ final class CrewInfoNoCrewView: UIView {
     private let noCrewLabel: UILabel = {
         let noCrewLabel = UILabel()
         noCrewLabel.text = "아직 참여중인\n크루가 없어요"
+        noCrewLabel.numberOfLines = 0
         noCrewLabel.font = UIFont.carmuFont.headline1
         noCrewLabel.textColor = UIColor.semantic.textPrimary
+        noCrewLabel.textAlignment = .center
         return noCrewLabel
     }()
 
     override init(frame: CGRect) {
         super.init(frame: frame)
+        self.layer.cornerRadius = 20
+        self.layer.masksToBounds = true
+        self.backgroundColor = UIColor.semantic.backgroundDefault
         setupUI()
-//        setupConstraints()
+        setupConstraints()
     }
 
     required init?(coder: NSCoder) {
@@ -54,12 +51,10 @@ final class CrewInfoNoCrewView: UIView {
 
     override func layoutSubviews() {
         super.layoutSubviews()
-        contentView.frame = bounds
     }
 
     private func setupUI() {
-        addSubview(contentView)
-        contentView.addSubview(noCrewStackView)
+        addSubview(noCrewStackView)
         noCrewStackView.addArrangedSubview(characterImageView)
         noCrewStackView.addArrangedSubview(noCrewLabel)
     }
@@ -77,5 +72,23 @@ final class CrewInfoNoCrewView: UIView {
         noCrewLabel.snp.makeConstraints { make in
             make.leading.trailing.equalToSuperview()
         }
+    }
+}
+
+// MARK: - 프리뷰 canvas 세팅
+import SwiftUI
+
+struct MyPageViewControllerRepresentable: UIViewControllerRepresentable {
+    typealias UIViewControllerType = MyPageViewController
+    func makeUIViewController(context: Context) -> MyPageViewController {
+        return MyPageViewController()
+    }
+    func updateUIViewController(_ uiViewController: MyPageViewController, context: Context) {
+    }
+}
+@available(iOS 13.0.0, *)
+struct MyPageViewPreview: PreviewProvider {
+    static var previews: some View {
+        MyPageViewControllerRepresentable()
     }
 }
